+++ spec/spec_helper.rb
@@ -1,8 +1,19 @@
 require 'rspec'
+require 'simplecov'
+require 'coveralls'
+
+SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter.new(
+  [
+    SimpleCov::Formatter::HTMLFormatter,
+    Coveralls::SimpleCov::Formatter
+  ]
+)
+SimpleCov.start
+
 require 'time_difference'
 
 RSpec.configure do |config|
-  config.formatter     = 'documentation'
+  config.formatter = 'documentation'
 
   # Configure Timezone for proper tests
   ENV['TZ'] = 'UTC'